refactor: Move ScreenIndexes enum in its own file
