+++ src/app/view/FlagshipUI/index.ts
@@ -1,2 +1,3 @@
 export * from '/app/view/FlagshipUI/hooks/useFlagshipUI'
 export * from '/app/view/FlagshipUI/FlagshipUIService'
+export * from '/app/view/FlagshipUI/ScreenIndexes'
